@@ -2,39 +2,38 @@
   "name": "feedit",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
+    "debug": "NODE_OPTIONS='--inspect' next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "db:push": "prisma db push --schema=./src/prisma/schema.prisma"
+    "lint": "next lint"
   },
   "dependencies": {
-    "@prisma/client": "^4.13.0",
-    "flowbite": "^1.6.5",
-    "flowbite-react": "^0.4.4",
-    "next": "13.3.1",
+    "@prisma/client": "5.3.1",
+    "@radix-ui/react-icons": "^1.3.0",
+    "@radix-ui/themes": "2.0.0",
+    "@vercel/kv": "^0.2.3",
+    "next": "13.5.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "rss-feed-emitter": "^3.2.3"
+    "rss-parser": "^3.13.0"
   },
   "devDependencies": {
-    "@types/node": "18.15.11",
-    "@types/react": "18.0.33",
-    "@types/react-dom": "18.0.11",
-    "autoprefixer": "^10.4.14",
-    "eslint": "8.37.0",
-    "eslint-config-next": "13.3.0",
-    "eslint-config-prettier": "^8.8.0",
-    "postcss": "^8.4.23",
-    "prettier": "^2.8.8",
-    "prettier-plugin-tailwindcss": "^0.2.8",
-    "prisma": "^4.13.0",
-    "tailwindcss": "^3.3.2",
-    "typescript": "5.0.3"
+    "@types/node": "20.8.0",
+    "@types/react": "18.2.24",
+    "@types/react-dom": "18.2.8",
+    "eslint": "^8.51.0",
+    "eslint-config-next": "^13.5.4",
+    "eslint-config-prettier": "^9.0.0",
+    "prettier": "^3.0.3",
+    "prisma": "5.3.1",
+    "typescript": "^5.2.2",
+    "vercel": "^32.4.1"
   },
   "engines": {
-    "node": "18"
+    "node": "20"
   },
   "prettier": {
     "semi": false,
