@@ -1,8 +1,8 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   experimental: {
-    appDir: true,
-  },
+    serverActions: true
+  }
 }
 
-module.exports = nextConfig
+export default nextConfig
