@@ -1,5 +1,11 @@
 'use client'
 
-export function Providers({ children }: { children: React.ReactNode }) {
-  return <>{children}</>
+import { Theme } from '@radix-ui/themes'
+
+export function Providers ({ children }: { children: React.ReactNode }) {
+  return (
+    <Theme appearance='dark' accentColor='amber' panelBackground='solid' radius='small'>
+      {children}
+    </Theme>
+  )
 }
