@@ -1,24 +1,13 @@
+import '@radix-ui/themes/styles.css'
 import { Providers } from './Providers'
 import { Tabs } from './Tabs'
-import './globals.css'
-import { Chivo_Mono, Montserrat } from 'next/font/google'
+import Login from './Login'
+import { Container } from '@radix-ui/themes'
 
 export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'Feedit',
+  description: 'your RSS feed reader',
 }
-
-const montserrat = Montserrat({
-  variable: '--font-montserrat',
-  display: 'swap',
-  subsets: ['latin'],
-})
-
-const chivoMono = Chivo_Mono({
-  variable: '--font-chivo-mono',
-  display: 'swap',
-  subsets: ['latin'],
-})
 
 export default function RootLayout({
   children,
@@ -26,13 +15,14 @@ export default function RootLayout({
   children: React.ReactNode
 }) {
   return (
-    <html lang="en" className={`${montserrat.variable} ${chivoMono.variable}`}>
-      <body className="min-h-screen">
+    <html lang="en" suppressHydrationWarning>
+      <body>
         <Providers>
-          <div className="max-w-xl mx-auto mt-48">
-            <Tabs />
+          <Container>
+            <Login />
             {children}
-          </div>
+            <Tabs />
+          </Container>
         </Providers>
       </body>
     </html>
