@@ -1,7 +1,14 @@
-export default function Home() {
+import { Grid, Heading } from '@radix-ui/themes'
+import { Suspense } from 'react'
+import ArticleList from './ArticleList'
+
+export default async function Home() {
   return (
-    <>
-      <h1>Home page</h1>
-    </>
+    <Grid gap={'6'}>
+      <Heading>Home page</Heading>
+      <Suspense fallback={<div>Loading...</div>}>
+        <ArticleList />
+      </Suspense>
+    </Grid>
   )
 }
