@@ -1,7 +1,9 @@
+import { Heading } from '@radix-ui/themes'
+
 export default function Profile() {
   return (
     <>
-      <h1>Profile Page</h1>
+      <Heading>Profile Page</Heading>
     </>
   )
 }
