@@ -1,7 +1,17 @@
-export default function Settings() {
+
+import { AddFeed } from './AddFeed'
+import { Suspense } from 'react'
+import FeedList from './FeedList'
+import { Grid, Heading } from '@radix-ui/themes'
+
+export default function Settings () {
   return (
-    <>
-      <h1>Settings Page</h1>
-    </>
+    <Grid gap='6'>
+      <Heading>Settings Page</Heading>
+      <AddFeed />
+      <Suspense fallback={<div>Loading...</div>}>
+        <FeedList />
+      </Suspense>
+    </Grid>
   )
 }
